requirements.txt
fastapi
uvicorn
python-dotenv
openai
scikit-learn
Levenshtein